+++ app/Http/Controllers/Admin/AdminUserController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\Admin;
 
+use Hash;
 use Illuminate\Http\Request;
 use DB;
 use App\Http\Requests;
@@ -18,9 +19,11 @@ class AdminUserController extends Controller
     public function index(Request $request)
     {
         //显示用户
+        check_admin_purview('0');
+        $get_session = session('Admin_Session');
         $search = $request->input('Search');
         $data = UsersModel::where('u_name','like','%'.$search.'%')->orderBy('id','asc')->paginate(25);
-        return view('Admin.User.list',['data'=>$data,'search'=>$search]);
+        return view('Admin.User.list',['data'=>$data,'search'=>$search,'get_session'=>$get_session]);
     }
     /**
      * 发送Ajax
@@ -30,6 +33,7 @@ class AdminUserController extends Controller
 
     public function Ajax(Request $request)
     {
+        check_admin_purview('0');
         $uname = $request->input('u_name');
         $data = DB::table('users')->where('u_name','=',$uname)->first();
         if($data){
@@ -47,7 +51,9 @@ class AdminUserController extends Controller
     public function create()
     {
         //
-        return view('Admin.User.add');
+        check_admin_purview('0');
+        $get_session = session('Admin_Session');
+        return view('Admin.User.add',['get_session'=>$get_session]);
     }
 
     /**
@@ -58,11 +64,16 @@ class AdminUserController extends Controller
      */
     public function store(Request $request)
     {
+        check_admin_purview('0');
         $data = $request->except('_token');
         $users = new UsersModel;
         $users -> u_status = $data['u_status'];
         $users -> u_name = $data['u_name'];
-        $users -> u_password = strtoupper(md5($data['u_password']));
+        $U_data = UsersModel::where('u_name','=',$data['u_name'])->first();
+        if (!$U_data == null) {
+          return back()->with('Error','用户名已存在');
+        }
+        $users -> u_password = Hash::make($data['u_password']);
         $users -> u_sex = $data['u_sex'];
         if (!$data['u_photo'] == null) {
           $users -> u_photo = $data['u_photo'];
@@ -72,7 +83,7 @@ class AdminUserController extends Controller
         $users -> u_phone = $data['u_phone'];
         $users -> u_email = $data['u_email'];
         $users -> u_money = $data['u_money'];
-        $users -> u_paypassword = strtoupper(md5($data['u_paypassword']));
+        $users -> u_paypassword = Hash::make($data['u_paypassword']);
         $users -> u_time = time();
         $db = $users -> save();
         if($db){
@@ -91,8 +102,10 @@ class AdminUserController extends Controller
      */
     public function edit($id)
     {
-       $data =  UsersModel::find($id);
-        return view('Admin.User.edit',['data'=>$data]);
+        check_admin_purview('0');
+        $get_session = session('Admin_Session');
+        $data =  UsersModel::find($id);
+        return view('Admin.User.edit',['data'=>$data,'get_session'=>$get_session]);
     }
 
     /**
@@ -104,16 +117,17 @@ class AdminUserController extends Controller
      */
     public function update(Request $request, $id)
     {
+        check_admin_purview('0');
         $data = $request->except('_token');
         $users = UsersModel::find($id);
         $users -> u_status = $data['u_status'];
-        $users -> u_password = strtoupper(md5($data['u_password']));
+        $users -> u_password = Hash::make($data['u_password']);
         $users -> u_sex = $data['u_sex'];
         $users -> u_photo = $data['u_photo'];
         $users -> u_phone = $data['u_phone'];
         $users -> u_email = $data['u_email'];
         $users -> u_money = $data['u_money'];
-        $users -> u_paypassword = strtoupper(md5($data['u_paypassword']));
+        $users -> u_paypassword = Hash::make($data['u_paypassword']);
         $db = $users -> save();
         if($db){
             return redirect('/admin/user/index')->with('Success','修改成功');
@@ -130,6 +144,7 @@ class AdminUserController extends Controller
      */
     public function destroy($id)
     {
+        check_admin_purview('0');
         $data = UsersModel::find($id);
         $db = $data -> delete();
         if($db){
@@ -143,9 +158,11 @@ class AdminUserController extends Controller
      */
     public function recycled()
     {
+        check_admin_purview('0');
+        $get_session = session('Admin_Session');
         $data = UsersModel::onlyTrashed()->paginate(25);
         //var_dump($date);
-        return view('Admin.User.recycled',['data'=>$data]);
+        return view('Admin.User.recycled',['data'=>$data,'get_session'=>$get_session]);
     }
     /**
      * 会员回收站管理
@@ -154,6 +171,7 @@ class AdminUserController extends Controller
      */
     public  function recover($id)
     {
+        check_admin_purview('0');
         $db = UsersModel::withTrashed()->where('id',$id)->restore();
         if($db){
             return redirect('/admin/user/recycled')->with('Success','恢复成功');
@@ -164,6 +182,7 @@ class AdminUserController extends Controller
 
     public function delete($id)
     {
+        check_admin_purview('0');
         $db = DB::table('users')->where('id','=',$id)->delete();
         if($db){
            return redirect('/admin/user/recycled')->with('Success','删除成功');
